Disable fullscreen in Previewer
This caused the answer buttons to be shown twice after 6975/6961.
Fullscreen didn't make sense in the previewer to me, so I removed it.

Fixes 7027
